requirements.txt
# REQUIRE PYTHON 3.11.x
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
numpy==1.26.3
requests==2.31.0
flask==3.0.2
python-dotenv==1.0.1
SpeechRecognition==3.10.1
PyAudio==0.2.14
openai==1.12.0
langchain-core>=0.3.58,<1.0.0
langchain-community>=0.0.21
langchain>=0.1.0
langsmith>=0.1.125,<0.4
langchain-openai==0.3.16
pyttsx3==2.90
flask-cors==4.0.0 